requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
apprise==1.7.3
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
base58==1.0.3
binaryornot==0.4.4
bitarray==2.9.2
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
cookiecutter==2.6.0
coolname==2.2.0
croniter==2.0.2
cryptography==42.0.5
cytoolz==0.12.3
dataclassy==0.11.1
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.6.1
docker==6.1.3
ecdsa==0.18.0
eip712==0.2.4
email_validator==2.1.1
eth-abi==4.2.1
eth-account==0.10.0
eth-ape==0.7.11
eth-bloom==3.0.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.4.0
eth-pydantic-types==0.1.0a5
eth-rlp==1.0.1
eth-tester==0.9.1b2
eth-typing==3.5.2
eth-utils==2.3.1
ethpm-types==0.6.7
evm-trace==0.1.2
evmchains==0.0.4
executing==2.0.1
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
giza-actions @ git+https://github.com/gizatechxyz/actions-sdk.git@5c5c9f3b3daefedd3c3c03a69312f3c2bd61ca04
giza-cli @ file:///Users/lancedavis/Documents/Giza/projects/giza-cli
giza-osiris==0.2.5
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.3
griffe==0.41.3
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
ijson==3.2.3
importlib_metadata==7.0.2
iniconfig==2.0.0
ipython==8.22.2
iso8601==2.1.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==27.2.0
lazyasd==0.1.4
lru-dict==1.2.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
morphys==1.0
mpmath==1.3.0
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.17.1
orjson==3.9.15
packaging==23.2
pandas==1.5.3
parsimonious==0.9.0
parso==0.8.3
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pluggy==1.4.0
polars==0.19.19
prefect==2.14.6
prefect-docker==0.4.5
prefect-gcp==0.5.4
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==4.3.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
pyethash==0.1.27
pyflipper==0.18
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
pytest==7.4.4
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-slugify==8.0.4
pytz==2023.4
pytzdata==2020.1
pyunormalize==15.1.0
PyYAML==6.0.1
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rlp==3.0.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safe-pysha3==1.0.4
semantic-version==2.10.0
semver==3.0.2
serial==0.0.97
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
toolz==0.12.1
tqdm==4.66.2
traitlets==5.14.1
trie==2.2.0
typer==0.9.0
types-pyasn1==0.5.0.20240301
types-python-dateutil==2.8.19.20240106
types-python-jose==3.3.4.20240106
types-requests==2.31.0.20240218
typing_extensions==4.10.0
tzlocal==5.2
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.13
web3==6.15.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
